+++ lib/entities/anki/anki_note.dart
@@ -2,11 +2,14 @@
 import 'dart:math';
 
 // Package imports:
+import 'package:da_kanji_mobile/entities/isar/isars.dart';
+import 'package:da_kanji_mobile/widgets/dictionary/dictionary_example_tab.dart';
 import 'package:database_builder/database_builder.dart';
 
 // Project imports:
 import 'package:da_kanji_mobile/entities/iso/iso_table.dart';
 import 'package:da_kanji_mobile/repositories/anki/anki_data.dart';
+import 'package:get_it/get_it.dart';
 
 /// Represents a DaKanji-style Anki note
 class AnkiNote{
@@ -101,15 +104,18 @@ class AnkiNote{
     this.deckName, JMdict entry,
     {
       required List<String> langsToInclude,
+      required bool includeExample,
       int translationsPerLang = 3
     }
   )
     : translations = {}
   {
 
-    for (var meaning in entry.meanings) {
+    for (String langCode in langsToInclude) {
 
-      if(!langsToInclude.contains(meaning.language!)) continue;
+      final meanings = entry.meanings.where((e) => e.language == langCode);
+      if(meanings.isEmpty) continue;
+      final meaning = meanings.first;
 
       translations.putIfAbsent(meaning.language!, () => []);
       List<String> langTrans = meaning.meanings
@@ -128,12 +134,60 @@ class AnkiNote{
 
     //audio = ;
 
-    //example = ;
-
     //exampleAudio = ;
 
     //image = ;
 
+  }
+
+  /// Searches the examples database for sentences and adds them to this card if
+  /// any are found
+  Future setExamplesFromDict(
+    JMdict entry,
+    {
+      required List<String> langsToInclude,
+      required int numberOfExamples,
+      required bool includeTranslations,
+    }
+  ) async {
+
+    List<ExampleSentence> examples = await searchExamples(
+      langsToInclude, entry.kanjis, entry.readings, entry.hiraganas,
+      numberOfExamples, GetIt.I<Isars>().examples.directory
+    );
+
+    if(examples.isEmpty) return;
+
+    // get the positions where the entry matches the example
+    final spans = getMatchSpans(entry, examples);
+
+    example = "";
+    for (var i = 0; i < min(numberOfExamples, examples.length); i++) {
+
+      if(examples.length > 1) example += "${i+1}. ";
+      
+      // set the japanese example
+      for (var span in spans[i]) {
+        example += examples[i].sentence.replaceRange(span.item1, span.item2,
+          "<b>${examples[i].sentence.substring(span.item1, span.item2)}</b>");
+      }
+      example += "<br>";
+
+      // add the translations
+      if(includeTranslations){
+        for (var langCode in langsToInclude) {
+
+          final translations = examples[i].translations
+            .where((e) => e.language == langCode);
+          
+          if(translations.isNotEmpty) example += "${translations.first.sentence}\n";
+
+        }
+      }
+
+      if(examples.length - 1 != i) example += "<br>";
+      
+    }
 
   }