+++ lib/widgets/dictionary/dictionary_example_tab.dart
@@ -67,27 +67,18 @@ class _DictionaryExampleTabState extends State<DictionaryExampleTab> {
       List<String> selectedLangs = 
         GetIt.I<Settings>().dictionary.selectedTranslationLanguages;
 
-      List<String> kanjiSplits = widget.entry!.kanjis.map((e) => 
-        GetIt.I<Mecab>().parse(e)
-          .where((e) => e.features.length > 6)
-          .map((e) => e.features[6]).toList()
-      )
-      .expand((e) => e)
-      .toList();
-
-      examplesSearch = (compute(searchExamples, Tuple7(
+      searchExamples(
         selectedLangs,
         widget.entry!.kanjis,
         widget.entry!.readings,
         widget.entry!.hiraganas,
-        kanjiSplits,
         limit,
         GetIt.I<Isars>().examples.directory
-      )).then((value) {
+      ).then((value) {
         examples = value;
-        matchSpans = getMatchSpans();
+        matchSpans = getMatchSpans(widget.entry!, examples);
         return examples;
-      }));
+      });
     }    
   }
   
@@ -149,8 +140,11 @@ class _DictionaryExampleTabState extends State<DictionaryExampleTab> {
     
   }
 
+}
+
   /// Get a list of lists of TextSpan that mark the entry's word in the example sentences
-  List<List<Tuple2<int, int>>> getMatchSpans(){
+  List<List<Tuple2<int, int>>> getMatchSpans(
+    JMdict entry, List<ExampleSentence> examples){
 
     List<List<Tuple2<int, int>>> matchSpans = [];
 
@@ -161,7 +155,7 @@ class _DictionaryExampleTabState extends State<DictionaryExampleTab> {
       String example = examples[e].sentence;
       List<TokenNode> parsedExample = GetIt.I<Mecab>().parse(example);
 
-      for (List<String> items in [widget.entry!.kanjis, widget.entry!.readings, widget.entry!.hiraganas]){
+      for (List<String> items in [entry.kanjis, entry.readings, entry.hiraganas]){
         for (int k = 0; k < items.length; k++){
           String item = items[k];
 
@@ -187,13 +181,13 @@ class _DictionaryExampleTabState extends State<DictionaryExampleTab> {
             continue;
           }
 
-          if(widget.entry!.kanjis.contains(parsedExample[i].features[6])){
+          if(entry.kanjis.contains(parsedExample[i].features[6])){
             matchSpans.last.add(currentSpan);
           }
-          else if(widget.entry!.readings.contains(parsedExample[i].features[6])){
+          else if(entry.readings.contains(parsedExample[i].features[6])){
             matchSpans.last.add(currentSpan);
           }
-          else if(widget.entry!.hiraganas.contains(parsedExample[i].features[6])){
+          else if(entry.hiraganas.contains(parsedExample[i].features[6])){
             matchSpans.last.add(currentSpan);
           }
         }
@@ -201,18 +195,16 @@ class _DictionaryExampleTabState extends State<DictionaryExampleTab> {
     }
     return matchSpans;
   }
-}
 
 /// Searches for examples in the database that contain JMEntry.
-List<ExampleSentence> searchExamples(Tuple7 query){
+List<ExampleSentence> _searchExamples(Tuple6 query){
 
   List<String> selectedLangs = query.item1;
   List<String> kanjis = query.item2;
   List<String> readings = query.item3;
   List<String> hiraganas = query.item4;
-  //List<String> kanjiSplits = query.item5;
-  int limit = query.item6;
-  String isarPath = query.item7;
+  int limit = query.item5;
+  String isarPath = query.item6;
   
   // find all examples in ISAR that cotain this words kanji
   Isar examplesIsar = Isar.openSync(
@@ -271,4 +263,27 @@ List<ExampleSentence> searchExamples(Tuple7 query){
   });
 
   return examples;
+}
+
+/// Searches for example sentences, same as `_searchExamples` but does so in 
+/// an isolate to prevent blocking the UI. Therefore, this method should be
+/// preferred over `_searchExamples`
+Future<List<ExampleSentence>> searchExamples(
+  List<String> selectedLangs,
+  List<String> kanjis,
+  List<String> readings,
+  List<String> hiraganas,
+  int limit,
+  String? directory
+){
+
+  return compute(_searchExamples, Tuple6(
+    selectedLangs,
+    kanjis,
+    readings,
+    hiraganas,
+    limit,
+    directory
+  ));
+
 }